create fsat report service and move logic for fsat assessments
