+++ src/app/report-rollup/fsat-report-rollup.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { FsatReportRollupService } from './fsat-report-rollup.service';
+
+describe('FsatReportRollupService', () => {
+  let service: FsatReportRollupService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(FsatReportRollupService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
