+++ src/app/report-rollup/report-rollup.service.ts
@@ -27,26 +27,22 @@ import { OpportunityCardsService, OpportunityCardData } from '../treasure-hunt/t
 import { OpportunityPaybackService } from '../treasure-hunt/treasure-hunt-report/opportunity-payback.service';
 import { PsatReportRollupService } from './psat-report-rollup.service';
 import { PhastReportRollupService } from './phast-report-rollup.service';
+import { FsatReportRollupService } from './fsat-report-rollup.service';
 
 
 @Injectable()
 export class ReportRollupService {
 
   reportAssessments: BehaviorSubject<Array<ReportItem>>;
-  fsatAssessments: BehaviorSubject<Array<ReportItem>>;
   ssmtAssessments: BehaviorSubject<Array<ReportItem>>;
   treasureHuntAssessments: BehaviorSubject<Array<ReportItem>>;
 
   assessmentsArray: Array<ReportItem>;
-  fsatArray: Array<ReportItem>;
   ssmtArray: Array<ReportItem>;
   treasureHuntArray: Array<ReportItem>;
 
 
 
-  selectedFsats: BehaviorSubject<Array<FsatCompare>>;
-  fsatResults: BehaviorSubject<Array<FsatResultsData>>;
-  allFsatResults: BehaviorSubject<Array<AllFsatResultsData>>;
 
   selectedSsmt: BehaviorSubject<Array<SsmtCompare>>;
   ssmtResults: BehaviorSubject<Array<SsmtResultsData>>;
@@ -56,7 +52,6 @@ export class ReportRollupService {
 
   calcsArray: Array<Calculator>;
   selectedCalcs: BehaviorSubject<Array<Calculator>>;
-  numFsats: number = 0;
   numSsmt: number = 0;
   numTreasureHunt: number = 0;
   showSummaryModal: BehaviorSubject<string>;
@@ -66,14 +61,14 @@ export class ReportRollupService {
     private assessmentDbService: AssessmentDbService,
     private settingsDbService: SettingsDbService,
     private calculatorDbService: CalculatorDbService,
-    private fsatService: FsatService,
     private ssmtService: SsmtService,
     private treasureHuntReportService: TreasureHuntReportService,
     private opportunitySummaryService: OpportunitySummaryService,
     private opportunityCardsService: OpportunityCardsService,
     private opportunityPaybackService: OpportunityPaybackService,
     private psatReportRollupService: PsatReportRollupService,
-    private phastReportRollupService: PhastReportRollupService
+    private phastReportRollupService: PhastReportRollupService,
+    private fsatReportRollupService: FsatReportRollupService
 
   ) {
     this.initSummary();
@@ -82,17 +77,15 @@ export class ReportRollupService {
   initSummary() {
     this.psatReportRollupService.initSummary();
     this.phastReportRollupService.initSummary();
+    this.fsatReportRollupService.initSummary();
 
     this.reportAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
-    this.fsatAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
+
     this.ssmtAssessments = new BehaviorSubject<Array<ReportItem>>(new Array<ReportItem>());
     this.treasureHuntAssessments = new BehaviorSubject<Array<ReportItem>>(Array<ReportItem>());
 
 
 
-    this.selectedFsats = new BehaviorSubject<Array<FsatCompare>>(new Array<FsatCompare>());
-    this.fsatResults = new BehaviorSubject<Array<FsatResultsData>>(new Array<FsatResultsData>());
-    this.allFsatResults = new BehaviorSubject<Array<AllFsatResultsData>>(new Array<AllFsatResultsData>());
 
     this.selectedSsmt = new BehaviorSubject<Array<SsmtCompare>>(new Array<SsmtCompare>());
     this.ssmtResults = new BehaviorSubject<Array<SsmtResultsData>>(new Array<SsmtResultsData>());
@@ -110,9 +103,7 @@ export class ReportRollupService {
     let settings = this.settingsDbService.getByAssessmentId(assessment);
     let tmpSettings = this.checkSettings(settings);
     this.assessmentsArray.push({ assessment: assessment, settings: tmpSettings });
-    if (assessment.fsat) {
-      this.fsatArray.push({ assessment: assessment, settings: tmpSettings });
-    } else if (assessment.ssmt) {
+    if (assessment.ssmt) {
       this.ssmtArray.push({
         assessment: assessment,
         settings: tmpSettings
@@ -131,7 +122,7 @@ export class ReportRollupService {
     this.assessmentsArray = new Array<ReportItem>();
     // this.phastArray = new Array<ReportItem>();
     // this.psatArray = new Array<ReportItem>();
-    this.fsatArray = new Array<ReportItem>();
+    // this.fsatArray = new Array<ReportItem>();
     this.ssmtArray = new Array<ReportItem>();
     this.treasureHuntArray = new Array<ReportItem>();
     this.calcsArray = new Array<Calculator>();
@@ -157,7 +148,8 @@ export class ReportRollupService {
     this.phastReportRollupService.phastAssessments.next(phastArray);
     let psatArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.psat != undefined });
     this.psatReportRollupService.psatAssessments.next(psatArray);
-    this.fsatAssessments.next(this.fsatArray);
+    let fsatArray: Array<ReportItem> = _.filter(this.assessmentsArray, (assessmentItem) => { return assessmentItem.assessment.fsat != undefined });
+    this.fsatReportRollupService.fsatAssessments.next(fsatArray);
     this.ssmtAssessments.next(this.ssmtArray);
     this.treasureHuntAssessments.next(this.treasureHuntArray);
     this.reportAssessments.next(this.assessmentsArray);
@@ -194,73 +186,7 @@ export class ReportRollupService {
 
 
 
-  //USED FOR FSAT SUMMARY
-  initFsatCompare(resultsArr: Array<AllFsatResultsData>) {
-    let tmpResults: Array<FsatCompare> = new Array<FsatCompare>();
-    resultsArr.forEach(result => {
-      let minCost = _.minBy(result.modificationResults, (result) => { return result.annualCost; });
-      let modIndex = _.findIndex(result.modificationResults, { annualCost: minCost.annualCost });
-      let fsatAssessments = this.fsatAssessments.value;
-      let assessmentIndex = _.findIndex(fsatAssessments, (val) => { return val.assessment.id === result.assessmentId; });
-      let item = fsatAssessments[assessmentIndex];
-      if (result.isBaseline) {
-        tmpResults.push({ baseline: item.assessment.fsat, modification: item.assessment.fsat, assessmentId: result.assessmentId, selectedIndex: -1, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-      } else {
-        tmpResults.push({ baseline: item.assessment.fsat, modification: item.assessment.fsat.modifications[modIndex].fsat, assessmentId: result.assessmentId, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-      }
-    });
-    this.selectedFsats.next(tmpResults);
-  }
 
-  updateSelectedFsats(item: ReportItem, modIndex: number) {
-    let tmpSelected = this.selectedFsats.value;
-    if (modIndex !== -1) {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.fsat, modification: item.assessment.fsat.modifications[modIndex].fsat, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    } else {
-      let selectedIndex = _.findIndex(tmpSelected, { assessmentId: item.assessment.id });
-      tmpSelected.splice(selectedIndex, 1, { baseline: item.assessment.fsat, modification: item.assessment.fsat, assessmentId: item.assessment.id, selectedIndex: modIndex, name: item.assessment.name, assessment: item.assessment, settings: item.settings });
-    }
-    this.selectedFsats.next(tmpSelected);
-  }
-
-  initFsatResultsArr(fsatArr: Array<ReportItem>) {
-    let tmpResultsArr = new Array<AllFsatResultsData>();
-    fsatArr.forEach(val => {
-      if (val.assessment.fsat.setupDone) {
-        let baselineResults = this.fsatService.getResults(JSON.parse(JSON.stringify(val.assessment.fsat)), true, val.settings);
-        if (val.assessment.fsat.modifications) {
-          if (val.assessment.fsat.modifications.length !== 0) {
-            let modResultsArr = new Array<FsatOutput>();
-            val.assessment.fsat.modifications.forEach(mod => {
-              let tmpResults: FsatOutput = this.fsatService.getResults(JSON.parse(JSON.stringify(mod.fsat)), false, val.settings);
-              modResultsArr.push(tmpResults);
-            });
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id });
-          } else {
-            let modResultsArr = new Array<FsatOutput>();
-            modResultsArr.push(baselineResults);
-            tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-          }
-        } else {
-          let modResultsArr = new Array<FsatOutput>();
-          modResultsArr.push(baselineResults);
-          tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modResultsArr, assessmentId: val.assessment.id, isBaseline: true });
-        }
-      }
-    });
-    this.allFsatResults.next(tmpResultsArr);
-  }
-
-  getFsatResultsFromSelected(selectedFsats: Array<FsatCompare>) {
-    let tmpResultsArr = new Array<FsatResultsData>();
-    selectedFsats.forEach(val => {
-      let baselineResults: FsatOutput = this.fsatService.getResults(JSON.parse(JSON.stringify(val.baseline)), true, val.settings);
-      let modificationResults: FsatOutput = this.fsatService.getResults(JSON.parse(JSON.stringify(val.modification)), false, val.settings);
-      tmpResultsArr.push({ baselineResults: baselineResults, modificationResults: modificationResults, assessmentId: val.assessmentId, name: val.name, modName: val.modification.name, baseline: val.baseline, modification: val.modification, settings: val.settings });
-    });
-    this.fsatResults.next(tmpResultsArr);
-  }
 
   //used for ssmt summary
   initSsmtCompare(resultsArr: Array<AllSsmtResultsData>) {
