+++ src/app/report-rollup/report-summary/fsat-summary/fsat-summary.component.html
@@ -2,7 +2,7 @@
   <div class="card-header">
       <h3>Fans</h3>
       <span>
-          <span>{{reportRollupService.numFsats}}</span>
+          <span>{{numFsats}}</span>
       </span>
   </div>
   <div class="card-body">
