+++ src/app/shared/print-options-menu/print-options-menu.component.ts
@@ -7,6 +7,7 @@ import { WindowRefService } from '../../indexedDb/window-ref.service';
 import { ReportRollupService } from '../../report-rollup/report-rollup.service';
 import { PsatReportRollupService } from '../../report-rollup/psat-report-rollup.service';
 import { PhastReportRollupService } from '../../report-rollup/phast-report-rollup.service';
+import { FsatReportRollupService } from '../../report-rollup/fsat-report-rollup.service';
 @Component({
   selector: 'app-print-options-menu',
   templateUrl: './print-options-menu.component.html',
@@ -26,7 +27,7 @@ export class PrintOptionsMenuComponent implements OnInit {
   showTHReportOptions: boolean = false;
   showWasteWaterOptions: boolean = false;
   constructor(private printOptionsMenuService: PrintOptionsMenuService, private windowRefService: WindowRefService, private reportRollupService: ReportRollupService,
-    private psatReportRollupService: PsatReportRollupService, private phastReportRollupService: PhastReportRollupService) { }
+    private psatReportRollupService: PsatReportRollupService, private phastReportRollupService: PhastReportRollupService, private fsatReportRollupService: FsatReportRollupService) { }
 
   ngOnInit() {
     this.setContext();
@@ -60,7 +61,7 @@ export class PrintOptionsMenuComponent implements OnInit {
     } else if (printContext == 'reportRollup') {
       this.showRollupReportOptions = true;
       this.showPsatReportOptions = (this.psatReportRollupService.psatAssessments.getValue().length != 0);
-      this.showFsatReportOptions = (this.reportRollupService.numFsats != 0);
+      this.showFsatReportOptions = (this.fsatReportRollupService.fsatAssessments.getValue().length != 0);
       this.showPhastReportOptions = (this.phastReportRollupService.phastAssessments.getValue().length != 0);
       this.showSsmtReportOptions = (this.reportRollupService.numSsmt != 0);
       this.showTHReportOptions = (this.reportRollupService.numTreasureHunt != 0);
